Refactor add packages
