+++ src/main/java/gray/GrayException.java
@@ -1,3 +1,5 @@
+package gray;
+
 public class GrayException extends Exception {
     public GrayException(String message) {
         super(message);
