+++ src/main/java/gray/Storage.java
@@ -1,3 +1,7 @@
+package gray;
+
+import gray.task.TaskList;
+
 import java.io.File;
 import java.io.IOException;
