+++ src/main/java/gray/command/AddTodoCommand.java
@@ -1,20 +1,22 @@
-import java.time.LocalDateTime;
+package gray.command;
 
-public class AddDeadlineCommand extends Command {
+import gray.Ui;
+import gray.task.TaskList;
+import gray.task.TodoTask;
+
+public class AddTodoCommand extends Command {
 
     private final String description;
-    private final LocalDateTime deadline;
 
-    public AddDeadlineCommand(String description, LocalDateTime deadline) {
+    public AddTodoCommand(String description) {
         this.description = description;
-        this.deadline = deadline;
     }
 
     @Override
     public void execute(Ui ui, TaskList tasks) {
-        DeadlineTask task = new DeadlineTask(description, deadline);
+        TodoTask task = new TodoTask(description);
         tasks.add(task);
-        ui.say(String.format("""
+        ui.say(String.format("""                
                 Got it. I've added this task:
                     %s
                 Now you have %d tasks in the list.""",
