+++ src/main/java/gray/command/Command.java
@@ -1,3 +1,9 @@
+package gray.command;
+
+import gray.GrayException;
+import gray.Ui;
+import gray.task.TaskList;
+
 public abstract class Command {
     public void execute(Ui ui, TaskList tasks) throws GrayException { }
     public boolean isExit() { return false; }
