+++ src/main/java/gray/command_factory/ByeCommandFactory.java
@@ -1,3 +1,9 @@
+package gray.command_factory;
+
+import gray.GrayException;
+import gray.command.ByeCommand;
+import gray.command.Command;
+
 public class ByeCommandFactory extends CommandFactory {
     @Override
     public Command parse(String text) throws GrayException {
