+++ src/main/java/gray/task/DeadlineTask.java
@@ -1,3 +1,5 @@
+package gray.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
