+++ src/main/java/gray/task/Task.java
@@ -1,3 +1,5 @@
+package gray.task;
+
 import java.io.Serializable;
 
 public class Task implements Serializable {
