+++ src/main/java/gray/task/TodoTask.java
@@ -1,3 +1,5 @@
+package gray.task;
+
 public class TodoTask extends Task {
 
     public TodoTask(String description) {
